Task Submission done

+++ app/Http/Controllers/CategoriesController.php
@@ -44,7 +44,9 @@ class CategoriesController extends Controller
      */
     public function store(StoreCategory $request)
     {
-        $this->middleware(['role:superadmin', 'role:admin']);
+        if (!auth('api')->user()->hasAnyRole(['admin', 'superadmin'])) {
+            return $this->ERROR('You dont have the permission to perform this action');
+        }
 
         $data = $request->validated();
 
@@ -94,7 +96,9 @@ class CategoriesController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->middleware(['role:superadmin', 'role:admin']);
+        if (!auth('api')->user()->hasAnyRole(['admin', 'superadmin'])) {
+            return $this->ERROR('You dont have the permission to perform this action');
+        }
 
         $data = $request->validate([
             'title' => 'required|unique:categories',
@@ -122,7 +126,9 @@ class CategoriesController extends Controller
      */
     public function destroy($id)
     {
-        $this->middleware(['role:superadmin', 'role:admin']);
+        if (!auth('api')->user()->hasAnyRole(['admin', 'superadmin'])) {
+            return $this->ERROR('You dont have the permission to perform this action');
+        }
 
         if (Category::find($id)->delete()) {
             return response()->json([
